cdk_tomcat_deployment/requirements.txt
aws-cdk-lib==2.180.0
constructs>=10.0.0,<11.0.0

black==25.1.0
mypy==1.15.0
isort==6.0.0
PyYAML==6.0.2
types-PyYAML==6.0.12.20241230
